pyproject.toml
[project]
name = "multimodal-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-groq>=0.3.7",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.5",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "uvicorn>=0.35.0",
]
